+++ shared/src/lib.rs
@@ -1,0 +1,2 @@
+
+pub static TEST_URI_STRING: &str = "quilt+s3://quilt-example#package=akarve/test_dest";